deduplicate some functions & types
