+++ src/components/menus/content-contextual-menu.jsx
@@ -30,7 +30,8 @@ import { DialogsId } from '../../utils/UIconstants';
 import { ContingencyListType, FilterType } from '../../utils/elementType';
 import { ElementType, FilterCreationDialog, PARAM_LANGUAGE, useSnackMessage } from '@gridsuite/commons-ui';
 import CommonContextualMenu from './common-contextual-menu';
-import { useDeferredFetch, useMultipleDeferredFetch } from '../../utils/custom-hooks';
+import useDeferredFetch from '../../hooks/useDeferredFetch';
+import useMultipleDeferredFetch from '../../hooks/useMultipleDeferredFetch';
 import MoveDialog from '../dialogs/move-dialog';
 import { useDownloadUtils } from '../utils/caseUtils';
 import ExportCaseDialog from '../dialogs/export-case-dialog';
