+++ src/utils/notificationType.ts
@@ -5,8 +5,8 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-export const notificationType = {
-    DELETE_DIRECTORY: 'DELETE_DIRECTORY',
-    ADD_DIRECTORY: 'ADD_DIRECTORY',
-    UPDATE_DIRECTORY: 'UPDATE_DIRECTORY',
-};
+export enum notificationType {
+    DELETE_DIRECTORY = 'DELETE_DIRECTORY',
+    ADD_DIRECTORY = 'ADD_DIRECTORY',
+    UPDATE_DIRECTORY = 'UPDATE_DIRECTORY',
+}
